2.0 rewrite
Completely rewrote the mod and the branding
Should now be compatible with EVERY mod ever. Multiversion support is now unified from 1.20.1 and all future releases (hopefully). This is specific to Fabric.


Way more efficient DamageSource detection
Removed unused imports and code
Debloated anything that wasn't nessecary for the mod to function


Explosion tilt was removed. It now uses the non-directional tilt instead. The old version implementation was buggy and more often than not would tilt in an arbitrary direction. Moving forward, the mod will just be a lightweight fix to MC-259212.